Squashed commit of the following:
commit dbea65e783ca0ec4b035a0535b989bdd79e8c388
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 16:29:52 2022 +0100

    Update

commit 79890870a67c755439a47ab1b6dfc28ba8f9f1e3
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 16:11:09 2022 +0100

    Update

commit b2e310d660fa6b451cce2ae3171cf4d027521c57
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 16:00:25 2022 +0100

    Update nltepop.cc

commit 2c8447c8ed322440b32e81ead3de1f57741a8e2a
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 15:58:34 2022 +0100

    Update update_grid.cc

commit e2a55e7c84f95db70b7d06c7448c352111108f10
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 15:54:11 2022 +0100

    Update nonthermal.cc

commit 7df443aab44387fb805ab9844afde00186a5e983
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 15:50:04 2022 +0100

    Update ci.yml

commit efdcb3341b6e3190dfb4493192baec2656320f39
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 15:49:24 2022 +0100

    Update CI

commit 27e3d0b3918ab4c2aabaa9fa10915acc02cfcf28
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 15:31:22 2022 +0100

    Update nltepop.cc

commit 699effe4332c88e2a961dc30c41bbb05ea6230ab
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 15:18:21 2022 +0100

    Update nltepop.cc

commit 94e4cededc2e60b0c6cdfe6cada81bbbebe1c72d
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 15:05:59 2022 +0100

    Update nltepop.cc

commit ea14a0eb6b2f830f6f4e0f393f397e5cf7962c81
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 14:56:59 2022 +0100

    Update nltepop.cc

commit 8622ac05b68459e73cf6aa00c19e51c5cc9f3f50
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 14:48:57 2022 +0100

    Update grid.cc

commit aa02f1bbdcc8188c8ca49839cc44707811442cbd
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 14:40:25 2022 +0100

    Update nltepop.cc

commit 1c19470dd8ce9a8d7b456394e40abc228c4e7c37
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 14:28:24 2022 +0100

    Update nltepop.cc

commit 756397250c709eb64aa09f2ff8706f9b9a8551b4
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 14:17:50 2022 +0100

    Update grid.cc

commit 906c52d591512af840ec8648d675a7d8345942fb
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 14:09:29 2022 +0100

    Update nltepop.cc

commit c337737a82a4e000b56151bf0e20b0fbf9b4d9ce
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 14:03:10 2022 +0100

    Update update_grid.cc

commit 01128e4cd8c53d912f592fd31b5ea8cc31c94e97
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 13:56:10 2022 +0100

    Update ci.yml

commit 0b4ee3212a1b45042a5b837b83498ced4d4f5cd4
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 13:33:27 2022 +0100

    Revert "Revert "Update nonthermal.cc""

    This reverts commit a0bd5a0a189807f9ff04b34dafd369879f959e2b.

commit 86bc9b641f76c0653b7694a293bb9d77a010b195
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 13:25:29 2022 +0100

    Update atomic.cc

commit a0bd5a0a189807f9ff04b34dafd369879f959e2b
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 13:18:22 2022 +0100

    Revert "Update nonthermal.cc"

    This reverts commit 8fa2d8dff7be4b753534d0ffa1d93a69d739b1b9.

commit 5e396743ac80b678935f52049586dd01e93ab289
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 13:11:35 2022 +0100

    Update nltepop.cc

commit 8fa2d8dff7be4b753534d0ffa1d93a69d739b1b9
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 13:03:50 2022 +0100

    Update nonthermal.cc

commit 710133547db0aaa65db21640537aa14424766823
Author: Luke Shingles <luke.shingles@gmail.com>
Date:   Wed Apr 13 13:01:46 2022 +0100

    Update nltepop.cc

+++ atomic.cc
@@ -66,7 +66,7 @@ __host__ __device__
 double get_nntot(int modelgridindex)
 // total ion (nuclei) density
 {
-  int nntot = 0.;
+  double nntot = 0.;
   for (int element = 0; element < get_nelements(); element++)
   {
     nntot += grid::get_elem_numberdens(modelgridindex, element);
